DVC20/requirements.txt
python3
pip3
pyserial
stress-ng
sensors
xrandr